将基本数据类型全部修改为包装数据类型

+++ src/main/java/xz/fzu/controller/ResumeController.java
@@ -60,7 +60,7 @@ public class ResumeController {
      */
     @RequestMapping(value = "/user/getresume", method = RequestMethod.POST)
     @ResponseBody
-    public ResponseVO<Resume> getResume(@RequestParam String token, @RequestParam int resumeId) throws InstanceNotExistException, TokenExpiredException {
+    public ResponseVO<Resume> getResume(@RequestParam String token, @RequestParam Long resumeId) throws InstanceNotExistException, TokenExpiredException {
 
         ResponseVO<Resume> responseVO = new ResponseVO<>();
         String userId = iUserService.verifyToken(token);
